Issue #37: Updated user model, adding new columns, and ordering previous columns a little better.
Updated Model_Auth_User, so that when saving, if the currently-logged in user is updating their own profile, we set "updated_profile" to 1, and if the password has changed, set "updated_password" to 1.  Added some better commenting.

Noticed that the Model_Auth_User I was updating was actually part of the Kohana Auth module, not the CL4 module - moved save updates to Model_CL4_User.

Set a check in Controller_Login after login, if user hasn't updated profile or password, add a message and redirect to profile page.  Added new messages to account messages file for this.